feature/add-binary-code-generator

+++ src/CodeGenerator/function_table.cpp
@@ -8,7 +8,7 @@ FunctionTable::~FunctionTable() {
     this->names.clear();
 }
 
-void FunctionTable::add(std::string name, int address) {
+void FunctionTable::add(std::string name, long address) {
     this->names.push_back(std::make_pair(name, address));
 }
 
@@ -16,7 +16,7 @@ std::string FunctionTable::getName(int index) {
     return this->names[index].first;
 }
 
-int FunctionTable::getAddress(std::string name) {
+long FunctionTable::getAddress(std::string name) {
     for (int i = 0; i < (int) this->names.size(); i++) {
         if (this->names[i].first == name) {
             return this->names[i].second;
